+++ PositiveToNegativeTimerDemo/TimePart.swift
@@ -1,0 +1,12 @@
+//
+//  TimePart.swift
+//  PositiveToNegativeTimerDemo
+//
+//  Created by Elliot Knight on 05/10/2022.
+//
+
+import Foundation
+
+struct TimeParts {
+	var seconds = 0
+}
